Really fix dithering and palette mapping again (?)
Do it by *not* eagerly replacing colors early; instead, remember which
pixels the Histogram+IColorMapper selected from a palette and toss the
error when dithering if the new color still matches.  This ensures that
"close" matches -- below the limit of the palette's precision -- will
not result in a lot of dithering, especially when the palette may have
been especially constructed for it!

Also, fix bug trashing the greyscale lookup tables.